+++ src/emmental/data.py
@@ -130,7 +130,7 @@ class EmmentalDataset(Dataset):
                 raise ValueError(f"Label {name} should be torch.Tensor.")
 
         if self.Y_dict is None:
-            self.Y_dict = {}
+            self.Y_dict = dict()
         self._update_dict(self.Y_dict, Y_dict)
 
     def remove_feature(self, feature_name: str) -> None:
@@ -164,29 +164,23 @@ def emmental_collate_fn(
     X_batch: defaultdict = defaultdict(list)
     Y_batch: defaultdict = defaultdict(list)
 
-    # Learnable batch should be a pair of dict, while non learnable batch is a dict
-    is_learnable = True if not isinstance(batch[0], dict) else False
-
-    if is_learnable:
-        for x_dict, y_dict in batch:
-            if isinstance(x_dict, dict) and isinstance(y_dict, dict):
-                for field_name, value in x_dict.items():
-                    if isinstance(value, list):
-                        X_batch[field_name] += value
-                    else:
-                        X_batch[field_name].append(value)
-                for label_name, value in y_dict.items():
-                    if isinstance(value, list):
-                        Y_batch[label_name] += value
-                    else:
-                        Y_batch[label_name].append(value)
-    else:
-        for x_dict in batch:  # type: ignore
-            for field_name, value in x_dict.items():  # type: ignore
-                if isinstance(value, list):
-                    X_batch[field_name] += value
-                else:
-                    X_batch[field_name].append(value)
+    for item in batch:
+        # Check if batch is (x_dict, y_dict) pair
+        if isinstance(item, dict):
+            x_dict = item
+            y_dict: Dict[str, Any] = dict()
+        else:
+            x_dict, y_dict = item
+        for field_name, value in x_dict.items():
+            if isinstance(value, list):
+                X_batch[field_name] += value
+            else:
+                X_batch[field_name].append(value)
+        for label_name, value in y_dict.items():
+            if isinstance(value, list):
+                Y_batch[label_name] += value
+            else:
+                Y_batch[label_name].append(value)
 
     field_names = copy.deepcopy(list(X_batch.keys()))
 
@@ -203,15 +197,14 @@ def emmental_collate_fn(
             if item_mask_tensor is not None:
                 X_batch[f"{field_name}_mask"] = item_mask_tensor
 
-    if is_learnable:
-        for label_name, values in Y_batch.items():
-            Y_batch[label_name] = list_to_tensor(
-                values,
-                min_len=Meta.config["data_config"]["min_data_len"],
-                max_len=Meta.config["data_config"]["max_data_len"],
-            )[0]
+    for label_name, values in Y_batch.items():
+        Y_batch[label_name] = list_to_tensor(
+            values,
+            min_len=Meta.config["data_config"]["min_data_len"],
+            max_len=Meta.config["data_config"]["max_data_len"],
+        )[0]
 
-    if is_learnable:
+    if len(Y_batch) != 0:
         return dict(X_batch), dict(Y_batch)
     else:
         return dict(X_batch)
@@ -245,7 +238,13 @@ class EmmentalDataLoader(DataLoader):
         **kwargs: Any,
     ) -> None:
         """Initialize EmmentalDataLoader."""
-        assert isinstance(dataset, EmmentalDataset)
+        assert isinstance(
+            dataset, EmmentalDataset
+        ), "dataset should inherent from EmmentalDataset."
+        assert isinstance(
+            task_to_label_dict, dict
+        ), "task_to_label_dict should be a dict."
+
         super().__init__(dataset, collate_fn=collate_fn, **kwargs)
 
         self.task_to_label_dict = task_to_label_dict
@@ -253,19 +252,20 @@ class EmmentalDataLoader(DataLoader):
         self.uid = dataset.uid
         self.split = split
         self.n_batches = n_batches
-        self.is_learnable = True if not isinstance(dataset[0], dict) else False
 
-        if self.is_learnable:
-            for task_name, label_names in task_to_label_dict.items():
-                if not isinstance(label_names, list):
-                    label_names = [label_names]  # type: ignore
-                unrecognized_labels = set(label_names) - set(
-                    dataset[0][1].keys()  # type: ignore
+        for task_name, label_names in task_to_label_dict.items():
+            if label_names is None:
+                continue
+            if not isinstance(label_names, list):
+                label_names = [label_names]  # type: ignore
+            if not isinstance(dataset[0], dict):
+                unrecognized_labels = set(label_names) - set(list(dataset[0][1].keys()))
+            else:
+                unrecognized_labels = set(label_names)
+            if len(unrecognized_labels) > 0:
+                msg = (
+                    f"Unrecognized Label {unrecognized_labels} of Task "
+                    f"{task_name} in dataset {dataset.name}."
                 )
-                if len(unrecognized_labels) > 0:
-                    msg = (
-                        f"Unrecognized Label {unrecognized_labels} of Task {task_name} "
-                        f"in dataset {dataset.name}."
-                    )
-                    logger.warn(msg)
-                    raise ValueError(msg)
+                logger.error(msg)
+                raise ValueError(msg)