+++ src/emmental/task.py
@@ -40,7 +40,7 @@ class EmmentalTask(object):
         output_func: Callable,
         scorer: Scorer,
         action_outputs: Optional[List[Union[Tuple[str, str], Tuple[str, int]]]] = None,
-        module_device: Dict[str, Union[int, str, torch.device]] = {},
+        module_device: Dict[str, Union[int, str, torch.device]] = dict(),
         weight: Union[float, int] = 1.0,
         require_prob_for_eval: bool = True,
         require_pred_for_eval: bool = True,
@@ -63,6 +63,10 @@ class EmmentalTask(object):
         self.module_device = {}
         for module_name in module_device.keys():
             if module_name not in self.module_pool:
+                logger.warning(
+                    f"Module {module_name} from module_device doesn't in module_pool, "
+                    "skip..."
+                )
                 continue
             if module_device[module_name] == -1:
                 self.module_device[module_name] = torch.device("cpu")