+++ src/emmental/utils/parse_args.py
@@ -160,6 +160,13 @@ def parse_args(parser: Optional[ArgumentParser] = None) -> ArgumentParser:
     )
 
     learner_config.add_argument(
+        "--skip_learned_data",
+        type=str2bool,
+        default=False,
+        help="Iterate through dataloader when steps or epochs learned is true",
+    )
+
+    learner_config.add_argument(
         "--train_split",
         nargs="+",
         type=str,
@@ -181,13 +188,6 @@ def parse_args(parser: Optional[ArgumentParser] = None) -> ArgumentParser:
         type=str,
         default=["test"],
         help="The split for testing",
-    )
-
-    learner_config.add_argument(
-        "--ignore_index",
-        type=nullable_int,
-        default=None,
-        help="The ignore index, uses for masking samples",
     )
 
     learner_config.add_argument(
@@ -947,10 +947,10 @@ def parse_args_to_config(args: Namespace) -> Dict[str, Any]:
             "n_epochs": args.n_epochs,
             "steps_learned": args.steps_learned,
             "n_steps": args.n_steps,
+            "skip_learned_data": args.skip_learned_data,
             "train_split": args.train_split,
             "valid_split": args.valid_split,
             "test_split": args.test_split,
-            "ignore_index": args.ignore_index,
             "online_eval": args.online_eval,
             "optimizer_config": {
                 "optimizer": args.optimizer,