+++ tests/test_model.py
@@ -3,10 +3,12 @@ import logging
 import shutil
 from functools import partial
 
+import pytest
 from torch import nn as nn
 from torch.nn import functional as F
 
 from emmental import EmmentalModel, EmmentalTask, Meta, Scorer, init
+from emmental.modules.identity_module import IdentityModule
 
 logger = logging.getLogger(__name__)
 
@@ -20,10 +22,8 @@ def test_model(caplog):
     Meta.reset()
     init(dirpath)
 
-    def ce_loss(module_name, immediate_output_dict, Y, active):
-        return F.cross_entropy(
-            immediate_output_dict[module_name][0][active], (Y.view(-1))[active]
-        )
+    def ce_loss(module_name, immediate_output_dict, Y):
+        return F.cross_entropy(immediate_output_dict[module_name][0], Y.view(-1))
 
     def output(module_name, immediate_output_dict):
         return F.softmax(immediate_output_dict[module_name][0], dim=1)
@@ -109,5 +109,86 @@ def test_model(caplog):
 
     model.add_tasks([task1, task2])
     assert model.task_names == set(["task_1", "task_2"])
+
+    shutil.rmtree(dirpath)
+
+
+def test_model_invalid_task(caplog):
+    """Unit test of model with invalid task."""
+    caplog.set_level(logging.INFO)
+
+    dirpath = "temp_test_model_with_invalid_task"
+
+    Meta.reset()
+    init(dirpath)
+
+    task_name = "task1"
+
+    task = EmmentalTask(
+        name=task_name,
+        module_pool=nn.ModuleDict(
+            {
+                "input_module0": IdentityModule(),
+                f"{task_name}_pred_head": IdentityModule(),
+            }
+        ),
+        task_flow=[
+            {
+                "name": "input1",
+                "module": "input_module0",
+                "inputs": [("_input_", "data")],
+            },
+            {
+                "name": f"{task_name}_pred_head",
+                "module": f"{task_name}_pred_head",
+                "inputs": [("input1", 0)],
+            },
+        ],
+        module_device={"input_module0": -1},
+        loss_func=None,
+        output_func=None,
+        action_outputs=None,
+        scorer=None,
+        require_prob_for_eval=False,
+        require_pred_for_eval=True,
+    )
+
+    task1 = EmmentalTask(
+        name=task_name,
+        module_pool=nn.ModuleDict(
+            {
+                "input_module0": IdentityModule(),
+                f"{task_name}_pred_head": IdentityModule(),
+            }
+        ),
+        task_flow=[
+            {
+                "name": "input1",
+                "module": "input_module0",
+                "inputs": [("_input_", "data")],
+            },
+            {
+                "name": f"{task_name}_pred_head",
+                "module": f"{task_name}_pred_head",
+                "inputs": [("input1", 0)],
+            },
+        ],
+        module_device={"input_module0": -1},
+        loss_func=None,
+        output_func=None,
+        action_outputs=None,
+        scorer=None,
+        require_prob_for_eval=False,
+        require_pred_for_eval=True,
+    )
+
+    model = EmmentalModel(name="test")
+    model.add_task(task)
+
+    with pytest.raises(ValueError):
+        model.add_task(task1)
+
+    with pytest.raises(ValueError):
+        model.add_task(task_name)
 
     shutil.rmtree(dirpath)