+++ tests/utils/test_parse_args.py
@@ -55,10 +55,10 @@ def test_parse_args(caplog):
             "n_epochs": 1,
             "steps_learned": 0,
             "n_steps": None,
+            "skip_learned_data": False,
             "train_split": ["train"],
             "valid_split": ["valid"],
             "test_split": ["test"],
-            "ignore_index": None,
             "online_eval": False,
             "optimizer_config": {
                 "optimizer": "adam",
@@ -241,10 +241,10 @@ def test_checkpoint_metric(caplog):
             "n_epochs": 1,
             "steps_learned": 0,
             "n_steps": None,
+            "skip_learned_data": False,
             "train_split": ["train"],
             "valid_split": ["valid"],
             "test_split": ["test"],
-            "ignore_index": None,
             "online_eval": False,
             "global_evaluation_metric_dict": None,
             "optimizer_config": {