+++ tests/utils/test_utils.py
@@ -1,12 +1,15 @@
 """Emmental utils unit tests."""
 import logging
+from functools import partial
 
 import numpy as np
+import pytest
 import torch
 
 from emmental.utils.utils import (
     array_to_numpy,
     construct_identifier,
+    convert_to_serializable_json,
     merge,
     move_to_device,
     nullable_float,
@@ -36,6 +39,9 @@ def test_prob_to_pred(caplog):
         )
         is True
     )
+
+    with pytest.raises(ValueError):
+        prob_to_pred(1.23)
 
 
 def test_pred_to_prob(caplog):
@@ -71,12 +77,17 @@ def test_array_to_numpy(caplog):
         is True
     )
     assert (
-        np.array_equal(torch.tensor([[1, 2], [3, 4]]), np.array([[1, 2], [3, 4]]))
+        np.array_equal(
+            array_to_numpy(torch.tensor([[1, 2], [3, 4]])), np.array([[1, 2], [3, 4]])
+        )
         is True
     )
     assert np.array_equal(
         array_to_numpy([[1, 2], [3, 4]], flatten=True), np.array([1, 2, 3, 4])
     )
+
+    with pytest.raises(ValueError):
+        array_to_numpy(1.23)
 
 
 def test_merge(caplog):
@@ -107,6 +118,9 @@ def test_str2bool(caplog):
     assert str2bool("n") is False
     assert str2bool("f") is False
     assert str2bool("0") is False
+
+    with pytest.raises(ValueError):
+        str2bool("o")
 
 
 def test_str2dict(caplog):
@@ -164,3 +178,28 @@ def test_random_string(caplog):
     assert len(random_string(10)) == 10
     assert len(random_string(5)) == 5
     assert random_string(5).islower() is True
+
+
+def test_convert_to_serializable_json(caplog):
+    """Unite test of convert_to_serializable_json."""
+    caplog.set_level(logging.INFO)
+
+    class abc:
+        a = 1
+
+    def cde(a, b):
+        return a, b
+
+    config = {
+        1: 1,
+        2: 2,
+        3: {4: cde, 5: [abc(), {1: 1, 2: partial(cde, 1)}]},
+        6: (abc(), 1),
+    }
+
+    assert convert_to_serializable_json(config) == {
+        1: 1,
+        2: 2,
+        3: {4: "Function: cde", 5: ["Class: abc", {1: 1, 2: "Function: cde"}]},
+        6: ("Class: abc", 1),
+    }